+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Unreleased Changes
 ------------------
 
+* Issue - Aws::S3 - Resolved an issue where the `Aws::S3::Client` would raise
+  an unhelpful error when credentials were not set. Now the client will
+  raise an `Aws::Errors::MissingCredentialsError` when an API operation is
+  called and credentials are not provided. Resolves
+  [GitHub issue #708](https://github.com/aws/aws-sdk-ruby/issues/708).
+
 * Issue - Aws::SharedCredentials - Now ignoring trailing whitespace from the
   access key id and secret access key values when loading the INI file.